Re-organize dependencies module
