+++ src/compiler/typescript_to_rust_redesign/azle_generate/src/ts_ast/mod.rs
@@ -1,0 +1,10 @@
+pub mod fn_decl;
+mod ts_array_type;
+mod ts_fn_or_constructor_type;
+mod ts_fn_type;
+mod ts_method_signature;
+mod ts_tuple_type;
+mod ts_type;
+pub mod ts_type_alias_decl;
+mod ts_type_lit;
+mod ts_type_ref;
